# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cloud-native Manufacturing Execution System API for phygital fabrication.\nProvides endpoints for order management, task/Kanban board operations,\nmachine control, telemetry, quality management, and real-time updates.",
        "title": "PravaraMES API",
        "contact": {
            "name": "PravaraMES Team",
            "url": "https://github.com/madfam-org/pravara-mes",
            "email": "support@pravara.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:4500",
    "basePath": "/v1",
    "paths": {
...
        "/telemetry": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns telemetry data with filtering by machine, metric type, and time range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telemetry"
                ],
                "summary": "List telemetry data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by machine ID (UUID)",
                        "name": "machine_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by metric type (temperature, pressure, etc.)",
                        "name": "metric_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter from time (RFC3339)",
                        "name": "from_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to time (RFC3339)",
                        "name": "to_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to return (default 100, max 1000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Telemetry data",
                        "schema": {
                            "$ref": "#/definitions/internal_api.TelemetryQueryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameter format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/telemetry/aggregated": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns aggregated telemetry data (min, max, avg) for a machine and metric",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telemetry"
                ],
                "summary": "Get aggregated telemetry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine ID (UUID)",
                        "name": "machine_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Metric type (temperature, pressure, etc.)",
                        "name": "metric_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From time (RFC3339), defaults to 24h ago",
                        "name": "from_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To time (RFC3339), defaults to now",
                        "name": "to_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Aggregation interval (hour, day), default hour",
                        "name": "interval",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aggregated telemetry data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing required parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/telemetry/batch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Inserts multiple telemetry records in a single request (max 1000)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telemetry"
                ],
                "summary": "Batch insert telemetry",
                "parameters": [
                    {
                        "description": "Batch of telemetry records",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.BatchTelemetryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Insertion confirmation with count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Validation error or empty batch",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/telemetry/latest": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the most recent telemetry record for a machine and metric type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telemetry"
                ],
                "summary": "Get latest telemetry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine ID (UUID)",
                        "name": "machine_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Metric type (temperature, pressure, etc.)",
                        "name": "metric_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Latest telemetry record",
                        "schema": {
                            "$ref": "#/definitions/types.Telemetry"
                        }
                    },
                    "400": {
                        "description": "Missing required parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "No telemetry found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.BatchTelemetryRequest": {
            "type": "object",
            "required": [
                "records"
            ],
            "properties": {
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.TelemetryRecord"
                    }
                }
            }
        },
...
        "internal_api.TelemetryQueryResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Telemetry"
                    }
                },
                "from_time": {
                    "type": "string"
                },
                "machine_id": {
                    "type": "string"
                },
                "metric_type": {
                    "type": "string"
                },
                "to_time": {
                    "type": "string"
                }
            }
        },
        "internal_api.TelemetryRecord": {
            "type": "object",
            "required": [
                "machine_id",
                "metric_type",
                "value"
            ],
            "properties": {
                "machine_id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "metric_type": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "types.Telemetry": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machine_id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "metric_type": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token from OIDC provider (format: \"Bearer \u003ctoken\u003e\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}